Scope down to utils.py only — keep nvfp4.py unchanged
Per reviewer feedback: nvfp4.py already uses "a" correctly and the
_parse_cuda_version refactor is unrelated cleanup. This PR now only
fixes _get_cuda_arch_list() in utils.py.
