+++ python/sglang/jit_kernel/nvfp4.py
@@ -8,7 +8,7 @@ from typing import TYPE_CHECKING, Optional, Tuple
 
 import torch
 
-from sglang.jit_kernel.utils import _parse_cuda_version, cache_once, load_jit
+from sglang.jit_kernel.utils import cache_once, load_jit
 
 if TYPE_CHECKING:
     from tvm_ffi.module import Module
@@ -69,6 +69,19 @@ def _nvfp4_cuda_flags() -> list[str]:
     ]
 
 
+def _parse_cuda_version() -> tuple[int, int]:
+    v = torch.version.cuda
+    if not v:
+        return (0, 0)
+    parts = v.split(".")
+    if len(parts) < 2:
+        return (0, 0)
+    try:
+        return int(parts[0]), int(parts[1])
+    except ValueError:
+        return (0, 0)
+
+
 def _get_nvfp4_cuda_arch_list() -> str:
     if not torch.cuda.is_available():
         raise RuntimeError("NVFP4 JIT kernels require CUDA.")
@@ -77,12 +90,22 @@ def _get_nvfp4_cuda_arch_list() -> str:
         raise RuntimeError(
             f"NVFP4 JIT kernels require compute capability >= 10.0, got {major}.{minor}."
         )
+    # NVFP4 kernels use architecture-family-specific instructions and must be
+    # compiled for `sm_*a` targets (e.g. sm_100a), not plain sm_100.
     archs = [f"{major}.{minor}a"]
     cuda_major, _cuda_minor = _parse_cuda_version()
     if cuda_major >= 13 and "10.3a" not in archs:
         # Match sgl-kernel AOT fatbin behavior on CUDA 13+ for Blackwell.
         archs.append("10.3a")
-    return " ".join(dict.fromkeys(archs))
+    # Preserve order while de-duplicating.
+    seen = set()
+    ordered_archs: list[str] = []
+    for arch in archs:
+        if arch in seen:
+            continue
+        seen.add(arch)
+        ordered_archs.append(arch)
+    return " ".join(ordered_archs)
 
 
 @contextmanager
